libchromeos: upgrade to syslog v6
We're trying to migrate from syslog v4, since it has a dependency
affected by RUSTSEC-2020-0071

BUG=b:239449434
TEST=CQ

Cq-Depend: chromium:3780192
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3781444
Tested-by: George Burgess <gbiv@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>

+++ libchromeos-rs/src/syslog.rs
@@ -33,7 +33,7 @@ pub fn get_syslog_logger(ident: String) -> Result<BasicLogger> {
         syslog::unix(Formatter3164 {
             facility: Facility::LOG_USER,
             hostname: None,
-            pid: getpid(),
+            pid: getpid() as u32,
             process: ident,
         })
         .map_err(Error::SyslogUnix)?,
